Revert "Add mockup auth section with API and token"
This reverts commit 1bbe2d1a0261fa6be6a45ec164d81413b311bde9.
