--- app/(auth)/login.tsx
@@ -1,94 +1,0 @@
-import { Link, useRouter } from 'expo-router';
-import React, { useState } from 'react';
-import { Pressable, StyleSheet, Text, TextInput, View } from 'react-native';
-import { z } from 'zod';
-import { colors } from '../../constants/colors';
-
-const loginSchema = z.object({
-  username: z.string().min(1),
-  password: z.string().min(1),
-});
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: colors.background,
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-  input: {
-    marginTop: 30,
-    paddingLeft: 30,
-    paddingRight: 30,
-    width: '100%',
-  },
-  button: {
-    marginTop: 30,
-    paddingTop: 10,
-    paddingBottom: 10,
-    width: '100%',
-    textAlign: 'center',
-    backgroundColor: colors.cardBackground,
-    fontSize: 24,
-  },
-  error: {
-    color: 'red',
-  },
-});
-
-export default function Login() {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const [errorMessage, setErrorMessage] = useState('');
-  const router = useRouter();
-
-  return (
-    <View>
-      <TextInput
-        style={styles.input}
-        placeholder="Username"
-        value={username}
-        onChangeText={setUsername}
-      />
-      <TextInput
-        style={styles.input}
-        placeholder="Password"
-        secureTextEntry
-        value={password}
-        onChangeText={setPassword}
-      />
-      {errorMessage ? <Text style={styles.error}>{errorMessage}</Text> : null}
-
-      <Pressable
-        onPress={async () => {
-          const validation = loginSchema.safeParse({ username, password });
-
-          if (!validation.success) {
-            setErrorMessage('Invalid username or password');
-            return;
-          }
-
-          const response = await fetch('/api/login', {
-            method: 'POST',
-            body: JSON.stringify({ username, password }),
-          });
-
-          const data = await response.json();
-
-          if (data.error) {
-            setErrorMessage(data.error);
-          }
-
-          if (data.success) {
-            router.push('/');
-          }
-        }}
-      >
-        <Text style={styles.button}>Login</Text>
-      </Pressable>
-      <Link style={styles.button} href="/">
-        Back
-      </Link>
-    </View>
-  );
-}
