+++ app/guests/[id].tsx
@@ -44,9 +44,8 @@ export default function Guests() {
         if (typeof id !== 'string') {
           return;
         }
-        const response = await fetch(`/api/${id}`);
+        const response = await fetch(`/${id}`);
         const fetchedGuest = await response.json();
-
         setGuest(fetchedGuest.guest);
       } catch (error) {
         console.error('Error fetching guest', error);
